Fix callsites

+++ src/System.Private.CoreLib/shared/System/SpanHelpers.Byte.cs
@@ -288,7 +288,7 @@ namespace System
                             }
 
                             // Find bitflag offset of first match and add to current offset
-                            return ((int)(byte*)offset) + BitOps.TrailingZeroCount(matches);
+                            return ((int)(byte*)offset) + (int)BitOps.TrailingZeroCount(matches);
                         } while ((byte*)nLength > (byte*)offset);
                     }
 
@@ -308,7 +308,7 @@ namespace System
                         else
                         {
                             // Find bitflag offset of first match and add to current offset
-                            return ((int)(byte*)offset) + BitOps.TrailingZeroCount(matches);
+                            return ((int)(byte*)offset) + (int)BitOps.TrailingZeroCount(matches);
                         }
                     }
 
@@ -340,7 +340,7 @@ namespace System
                         }
 
                         // Find bitflag offset of first match and add to current offset
-                        return ((int)(byte*)offset) + BitOps.TrailingZeroCount(matches);
+                        return ((int)(byte*)offset) + (int)BitOps.TrailingZeroCount(matches);
                     }
 
                     if ((int)(byte*)offset < length)
@@ -652,7 +652,7 @@ namespace System
                             }
 
                             // Find bitflag offset of first match and add to current offset
-                            return ((int)(byte*)offset) + BitOps.TrailingZeroCount(matches);
+                            return ((int)(byte*)offset) + (int)BitOps.TrailingZeroCount(matches);
                         } while ((byte*)nLength > (byte*)offset);
                     }
 
@@ -674,7 +674,7 @@ namespace System
                         else
                         {
                             // Find bitflag offset of first match and add to current offset
-                            return ((int)(byte*)offset) + BitOps.TrailingZeroCount(matches);
+                            return ((int)(byte*)offset) + (int)BitOps.TrailingZeroCount(matches);
                         }
                     }
 
@@ -708,7 +708,7 @@ namespace System
                         }
 
                         // Find bitflag offset of first match and add to current offset
-                        return ((int)(byte*)offset) + BitOps.TrailingZeroCount(matches);
+                        return ((int)(byte*)offset) + (int)BitOps.TrailingZeroCount(matches);
                     }
 
                     if ((int)(byte*)offset < length)
@@ -888,7 +888,7 @@ namespace System
                             }
 
                             // Find bitflag offset of first match and add to current offset
-                            return ((int)(byte*)offset) + BitOps.TrailingZeroCount(matches);
+                            return ((int)(byte*)offset) + (int)BitOps.TrailingZeroCount(matches);
                         } while ((byte*)nLength > (byte*)offset);
                     }
 
@@ -912,7 +912,7 @@ namespace System
                         else
                         {
                             // Find bitflag offset of first match and add to current offset
-                            return ((int)(byte*)offset) + BitOps.TrailingZeroCount(matches);
+                            return ((int)(byte*)offset) + (int)BitOps.TrailingZeroCount(matches);
                         }
                     }
 
@@ -948,7 +948,7 @@ namespace System
                         }
 
                         // Find bitflag offset of first match and add to current offset
-                        return ((int)(byte*)offset) + BitOps.TrailingZeroCount(matches);
+                        return ((int)(byte*)offset) + (int)BitOps.TrailingZeroCount(matches);
                     }
 
                     if ((int)(byte*)offset < length)
